Strip BOM from text detected as UTF-16
The name returned from cchardet includes the byte-order suffix, but
for Python codecs that means do not handle the BOM. Switching to
plain 'UTF-16' ensures the BOM is stripped.

Note that cchardet does not detect an encoding for UTF-16 files
without a BOM, but the code ensures the mapping only happens when
the expected BOM is actually present.